+++ packages/sdk-react/CHANGES.md
@@ -1,5 +1,17 @@
 fusionauth-react-sdk Changes
 
+Changes in 2.4.3
+
+- Correctly address issues with config and `isLoggedIn` in [169](https://github.com/FusionAuth/fusionauth-javascript-sdk/pull/169)
+
+Changes in 2.4.2
+
+- Fix a bug where the token is not refreshed in React-SDK.
+
+Changes in 2.4.1
+
+- Fix the bug where the user is logged out on a successful token refresh. Resolved in PR [161](https://github.com/FusionAuth/fusionauth-javascript-sdk/pull/161).
+
 Changes in 2.4.0
 
 - Include the response status code in the error passed to [onAutoRefreshFailure](https://github.com/FusionAuth/fusionauth-javascript-sdk/blob/main/packages/sdk-react/docs/interfaces/providers_FusionAuthProviderConfig.FusionAuthProviderConfig.md#onautorefreshfailure). See issue [#151](https://github.com/FusionAuth/fusionauth-javascript-sdk/issues/151).